Code structure refactoring.
